# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ABNovaMart APIs",
    "version": "1.0.0",
    "description": "API Documentation for the ABNovaMart Backend Server"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/user": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users",
        "description": "Retrieves all users with pagination (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer"
                        },
                        "currentLimit": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "totalCount": {
                          "type": "integer"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "hasPreviousPage": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "message": "Users fetched successfully",
                  "users": [
                    {
                      "_id": "507f1f77bcf86cd799439011",
                      "email": "john.doe@example.com",
                      "firstName": "John",
                      "lastName": "Doe",
                      "phone": "1234567890",
                      "role": "USER"
                    },
                    {
                      "_id": "507f1f77bcf86cd799439012",
                      "email": "jane.doe@example.com",
                      "firstName": "Jane",
                      "lastName": "Doe",
                      "phone": "0987654321",
                      "role": "ADMIN"
                    }
                  ],
                  "pagination": {
                    "currentPage": 1,
                    "currentLimit": 10,
                    "totalPages": 2,
                    "totalCount": 15,
                    "hasNextPage": true,
                    "hasPreviousPage": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Access denied. Admin privileges required."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get current user's cart",
        "description": "Retrieves the cart for the currently authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": { "type": "object" },
                        "items": {
                          "type": "array",
                          "items": { "type": "object" }
                        },
                        "summary": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Clear current user's cart",
        "description": "Removes all items from the cart for the currently authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart cleared successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/product-categories": {
      "get": {
        "tags": ["Product Categories"],
        "summary": "Get all product categories",
        "description": "Retrieves all product categories with pagination and optional search (accessible to all users)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term to filter categories by name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product categories fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "totalCategories": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "hasPrevPage": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "message": "Product categories fetched successfully",
                  "categories": [
                    {
                      "_id": "507f1f77bcf86cd799439011",
                      "name": "Electronics",
                      "slug": "electronics",
                      "parentId": null,
                      "created_at": "2024-01-01T00:00:00.000Z",
                      "updated_at": "2024-01-01T00:00:00.000Z"
                    }
                  ],
                  "pagination": {
                    "currentPage": 1,
                    "totalPages": 1,
                    "totalCategories": 1,
                    "limit": 10,
                    "hasNextPage": false,
                    "hasPrevPage": false
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Product Categories"],
        "summary": "Create a new product category",
        "description": "Creates a new product category (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "slug"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Category name"
                  },
                  "slug": {
                    "type": "string",
                    "description": "URL-friendly slug for the category"
                  },
                  "parentId": {
                    "type": "string",
                    "description": "Parent category ID for nested categories (optional)"
                  }
                }
              },
              "example": {
                "name": "Electronics",
                "slug": "electronics",
                "parentId": null
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                },
                "example": {
                  "success": true,
                  "message": "Product category created successfully",
                  "data": {
                    "_id": "507f1f77bcf86cd799439011",
                    "name": "Electronics",
                    "slug": "electronics",
                    "parentId": null,
                    "created_at": "2024-01-01T00:00:00.000Z",
                    "updated_at": "2024-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Validation Error",
                  "errors": ["Name is required", "Slug is required"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Access denied. Admin privileges required."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get all orders (Admin)",
        "description": "Retrieves all orders with details (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All orders fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OrderWithDetails"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Access denied. Admin privileges required."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/addresses": {
      "get": {
        "tags": ["Addresses"],
        "summary": "Get user's addresses",
        "description": "Retrieves all addresses for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Addresses fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Address"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User's unique identifier"
          },
          "id": {
            "type": "string",
            "description": "User's unique identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "phone": {
            "type": "string",
            "description": "User's phone number"
          },
          "role": {
            "type": "string",
            "enum": ["ADMIN", "USER"],
            "description": "User's role"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Category's unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the category"
          },
          "parentId": {
            "type": "string",
            "description": "Parent category ID for nested categories"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the category was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the category was last updated"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Product's unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "price": {
            "type": "number",
            "description": "Product price"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit"
          },
          "stock": {
            "type": "integer",
            "description": "Available stock"
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Product images URLs"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Product tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the product was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the product was last updated"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Address's unique identifier"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "fullName": {
            "type": "string",
            "description": "Full name for the address"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "addressLine": {
            "type": "string",
            "description": "Address line"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "description": "State"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether this is the default address"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the address was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the address was last updated"
          }
        }
      },
...
      "OrderWithDetails": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "totalAmount": {
            "type": "number"
          },
          "paymentStatus": {
            "type": "string",
            "enum": ["PENDING", "PAID", "FAILED"]
          },
          "paymentIntentId": {
            "type": "string"
          },
          "orderStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          },
          "shippingAddressId": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "OrderItems": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "orderId": {
                  "type": "string"
                },
                "productId": {
                  "$ref": "#/components/schemas/Product"
                },
                "quantity": {
                  "type": "integer"
                },
                "price": {
                  "type": "number"
                }
              }
            }
          },
          "Address": {
            "$ref": "#/components/schemas/Address"
          },
          "User": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    }
  }
}